PDOK-13906 lighttpd config include for flexibility
We need lighttpd to pass extra environment variables to the mapserv
FastCGI process so it can read data directly out of S3 using GDAL's
vsis3 virtual filesystem.

An include is probably the most flexible solution: if we (or anyone
else) needs more environment variables to be passed to mapserv, they can
simply be added in a config file.